pyproject.toml
[project]
name = "thumbnail-adder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.9.3",
    "apscheduler==3.10.4",
    "beanie==1.30.0",
    "kurigram==2.2.9",
    "motor>=3.7.1",
    "pyromod==3.1.6",
    "python-dotenv==1.0.1",
    "reloadium>=1.5.1",
    "tgcrypto==1.2.3",
    "uvloop>=0.21.0 ; sys_platform != 'win32'",
]
